more reorganisation of main.js
