+++ src/lib/verify-assets-exist.js
